midterm/logistics/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.2.2
django-ninja-crud==0.6.1
django-ninja-extra==0.21.2
django-ninja-jwt==5.3.2
dnspython==2.6.1
email_validator==2.2.0
gunicorn==22.0.0
idna==3.7
injector==0.22.0
kombu==5.4.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
qrcode==7.4.2
redis==5.0.8
redis-cli==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
